feat/ add msme pin location
