+++ src/components/modals/AddMSMEModal.tsx
@@ -24,7 +24,7 @@ import { useMSMEContext } from "@/contexts/MSMEContext";
 import { cn } from "@/lib/utils";
 import { LocationSelect } from "@/components/forms/LocationSelect";
 // import { Map } from "@vis.gl/react-google-maps";
-import Map from "@/components/map/Map";
+import { GoogleMap, Marker, useJsApiLoader } from "@react-google-maps/api";
 interface AddMSMEModalProps {
   isOpen: boolean;
   onClose: () => void;
@@ -46,6 +46,13 @@ export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
   const [yearEstablished, setYearEstablished] = useState("");
   const [dtiNumber, setDTINumber] = useState("");
   const [sectorId, setSectorId] = useState<number | null>(null);
+  const [latitude, setLatitude] = useState<number | null>(null);
+  const [longitude, setLongitude] = useState<number | null>(null);
+  const [marker, setMarker] = useState<{ lat: number; lng: number } | null>(
+    null,
+  );
+  const [productGallery, setProductGallery] = useState<string[]>([]);
+  const [majorProductLines, setMajorProductLines] = useState<string[]>([]);
 
   // Generate years for select (from 1900 to current year)
   const currentYear = new Date().getFullYear();
@@ -77,6 +84,10 @@ export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
     return Object.keys(newErrors).length === 0;
   };
 
+  const { isLoaded } = useJsApiLoader({
+    googleMapsApiKey: process.env.NEXT_PUBLIC_GOOGLE_MAPS_API_KEY!,
+  });
+
   const handleSubmit = async (e: React.FormEvent) => {
     e.preventDefault();
     if (!validateForm()) return;
@@ -98,10 +109,10 @@ export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
         dti_number: parseInt(dtiNumber),
         sectorId,
         createdAt: new Date(),
-        productGallery: [],
-        majorProductLines: [],
-        longitude: 0,
-        latitude: 0,
+        productGallery,
+        majorProductLines,
+        longitude,
+        latitude,
       });
 
       onClose();
@@ -119,6 +130,10 @@ export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
       setDTINumber("");
       setSectorId(null);
       setErrors({});
+      setLatitude(null);
+      setLongitude(null);
+      setProductGallery(["sample.jpg"]);
+      setMajorProductLines(["sample.jpg"]);
     } catch (error) {
       console.error("Error adding MSME:", error);
     } finally {
@@ -128,6 +143,24 @@ export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
 
   const handleSectorChange = (value: string) => {
     setSectorId(Number(value));
+  };
+
+  const defaultMapCenter = {
+    // Default latitude and longitude for Iloilo City, Philippines
+    lat: 10.7202,
+    lng: 122.5621,
+  };
+
+  const handleMapClick = async (event: google.maps.MapMouseEvent) => {
+    if (event.latLng) {
+      const newMarker = {
+        lat: event.latLng.lat(),
+        lng: event.latLng.lng(),
+      };
+      setMarker(newMarker);
+      setLatitude(newMarker.lat);
+      setLongitude(newMarker.lng);
+    }
   };
 
   return (
@@ -355,9 +388,23 @@ export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
                 Pin Location <span className="text-red-500">*</span>
               </Label>
 
-              <div className="container mx-auto p-4">
-                <Map />
-              </div>
+              {!isLoaded ? (
+                <div>Loading...</div>
+              ) : (
+                <div className="container mx-auto p-4">
+                  <GoogleMap
+                    mapContainerStyle={{
+                      width: "100%",
+                      height: "400px",
+                    }}
+                    center={defaultMapCenter}
+                    zoom={12}
+                    onClick={handleMapClick}
+                  >
+                    {marker && <Marker position={marker} />}
+                  </GoogleMap>
+                </div>
+              )}
             </div>
           </div>
           <div className="flex items-center justify-end gap-4">
