+++ src/types/MSME.ts
@@ -13,8 +13,8 @@ export type MSME = {
   provinceAddress: string;
   cityMunicipalityAddress: string;
   barangayAddress: string;
-  longitude: number;
-  latitude: number;
+  longitude: number | null;
+  latitude: number | null;
   yearEstablished: number;
   dti_number: number;
   sectorId: number;
